refactor(authz): the link-target probe moves beside the link scope rules
rowscope.go crossed the 500-line cap. linkscope.go already holds the
activity_link visibility rules, and the write-side question — may I point a
link at this record — belongs with the two read-side ones rather than among
the generic single-row probes.
