+++ backend/internal/platform/auth/rowscope.go
@@ -358,66 +358,6 @@ func EnsureVisibleForSubjectRights(ctx context.Context, tx pgx.Tx, table string,
 	return nil
 }
 
-// EnsureLinkTarget verifies an activity link's target row exists AND is
-// visible to the caller — an explicit row-scope probe, because the FK that
-// would otherwise catch a bad id is checked as the table owner and carries
-// no row scope at all: it accepts any id that exists, so without this a
-// guessed foreign UUID would persist a link to a record the caller may not
-// read. A link to an archived record is equally refused: the link would
-// outlive the row it names.
-//
-// It HOLDS the row it answers about, and that is what separates it from
-// EnsureVisibleLive. Every caller writes the reference afterwards, in the same
-// transaction: probe, then insert. Read unlocked, an archive committing in
-// between makes the answer stale before it is used, and the reference lands on
-// a record that is no longer live — the TOCTOU storekit.LockRow's own doc warns
-// against, on the target side. Under READ COMMITTED the lock is also the
-// re-check: a waiter wakes on the archived row and the liveness filter refuses
-// it, rather than acting on what it read before waiting.
-//
-// FOR SHARE, and the pairing is the point. It conflicts with the archive, which
-// UPDATEs the row, while two references onto one record do not conflict and have
-// no reason to queue behind each other — a person mid-ingest is referenced by
-// every message captured for them. The same pairing the activity_link trigger
-// takes on the activity it is about (migration 1788000100).
-//
-// Taken in the probe rather than at each write, unlike LockSubjectLive next door
-// in writescope.go. That lock is exclusive and would add an edge to two dozen
-// documented lock orders at once; this one is shared, so it orders nothing
-// against another reference and only ever waits for a writer of the very row
-// being referenced. Against that, forty call sites would each have to remember,
-// and the one that forgot would look exactly like the thirty-nine that did not.
-//
-// The few read paths that call this for its existence half — a contract listing
-// under an organization, a relationship read under its anchor — pay for it too:
-// they hold that one anchor for the length of a short read, which delays an
-// archive of it and blocks nothing else. That is the price of the probe having
-// one meaning.
-func EnsureLinkTarget(ctx context.Context, tx pgx.Tx, table string, id ids.UUID) error {
-	var args []any
-	arg := func(v any) int { args = append(args, v); return len(args) }
-	idPos := arg(id)
-
-	clause, err := ScopeClauseFor(ctx, table, "", arg)
-	if err != nil {
-		return err
-	}
-	q := fmt.Sprintf(`SELECT 1 FROM %s WHERE id = $%d AND archived_at IS NULL`, table, idPos)
-	if clause != "" {
-		q += " AND " + clause
-	}
-	q += " FOR SHARE"
-
-	var held int
-	if err := tx.QueryRow(ctx, q, args...).Scan(&held); err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return apperrors.ErrNotFound
-		}
-		return err
-	}
-	return nil
-}
-
 // VisibleTo probes whether one row passes the caller's row scope WITHOUT
 // erroring — for the dedupe pre-checks, which must answer 409 either way
 // but may only disclose the existing row's id when the caller could read
